+++ ehrest/src/main/java/com/everhomes/rest/talent/TalentExperienceConditionEnum.java
@@ -20,14 +20,14 @@ public enum TalentExperienceConditionEnum {
 		this.code = code;
 	}
 
-	public Byte getCode() {
+	public byte getCode() {
 		return code;
 	}
 
 	public static TalentExperienceConditionEnum fromCode(Byte code) {
 		if (code != null) {
 			for (TalentExperienceConditionEnum talentExperienceConditionEnum : TalentExperienceConditionEnum.values()) {
-				if (talentExperienceConditionEnum.getCode().byteValue() == code.byteValue()) {
+				if (talentExperienceConditionEnum.getCode() == code.byteValue()) {
 					return talentExperienceConditionEnum;
 				}
 			}
